+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.0.3 (XXX XX, 20XX)
+- `initData` or `initRows` will be handled after grid initialized.
+- Fixed a problem that numeric zero value cannot be set by using setCtrlValue method.
+- Added `readonly` column type for `bootstrap4` and `bulma` UI framework.
+
 ## 2.0.2 (May 8, 2019)
 - Fixed a problem on generating unusable AppendGrid.js file.